2/10 22:00 수정본

+++ src/main/java/com/example/team5_be/habit/domain/dto/HabitCreateRequestDTO.java
@@ -8,8 +8,8 @@ import lombok.*;
 @AllArgsConstructor
 @Builder
 public class HabitCreateRequestDTO {
-    private Long tagId;
-    private Long styleId;
+    private Integer tagId;
+    private Integer styleId;
 
     private String habitName;
     private String habitDefinition;
